moved sound files out of sound folder
